Ensure LDAP users can get activated with any registration setting
LDAP users get an auth_source_id attached to them during registration,
since they are effectively not self-registrating in the system,
the register service should allow them with any self_registration setting.

At the same time, the refactoring ensure that even an LDAP user that logs in for the first
time is being redirected to home to get the welcome tour.